requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0
requests>=2.31,<3.0
pydantic>=2.6,<3.0
PyYAML>=6.0,<7.0
lightgbm>=4.3,<5.0
scikit-learn>=1.4,<2.0
akshare>=1.15
curl-cffi>=0.7
tushare>=1.4
urllib3>=2.0,<3.0
